Add "Authentication" on the backend to check for correct user
This is not real authentication

+++ routes/ideas.js
@@ -42,18 +42,29 @@ router.post('/', async (req, res) => {
 // Update idea
 router.put('/:id', async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+    // Match the usernames, this is not a substitute for validation
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      { new: true },
-    );
-    res.json({ success: true, data: updatedIdea });
+        { new: true },
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+
+    // Usernames do not match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorised to update this resource',
+    });
   } catch (error) {
+    // Username does not match
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
   }
@@ -62,8 +73,18 @@ router.put('/:id', async (req, res) => {
 // Delete idea
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+    // Match the usernames, this is not a substitute for validation
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+
+    // Username does not match
+    res.status(403).json({
+      success: false,
+      error: 'You are not authorised to delete this resource',
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong' });
